Fix Dockerfile to bind to all container interfaces
When running a service in a container, it's by default bound to
all interfaces. The publish option in 'docker run' controls how
the port is exposed to the host. e.g

docker run -p 127.0.0.1:48620:48620 will publish the port
only to the host's loopback address.

docker run -p 48620:48620 will publish the port to all interfaces.

Hence, fix the docker build file to always bind to 0.0.0.0
inside the container.

Signed-off-by: Sabari Kumar Murugesan <smurugesan@vmware.com>