Remove obsolete files
The stdeb file was added as a way to generate a debian package
from the binary output of this build using the stdeb python library.
Unfortunately, stdeb does not expose enough parameters to generate
a good package.

This patch removes that stdeb and other references since we no longer attempt
to generate a *.deb from the output.

Signed-off-by: Eric Brown <browne@vmware.com>