Avoid 500 error on bad decrypt
Currently when decrypting a value using the credential API that
is not a valid encrypted string, it throws an error 500 due to
an AnsibleError exception.

Column should gracefully handle bad data such as an invalid value.

This change catches the exception and then gives a 400 response
to the user instead of a 500. The error message also states that
the value was invalid.

Fixes Issue #168