Handle ValueError in JSON schema validator
A uncaught ValueError exception occurs in the validator decorator
if non-JSON body is received in the request.

This patch catches the potential ValueError and if occurs, passes
the data to the validate_payload without doing a json.load.

This will result in the JSON validator working properly again and
avoiding an ugly 500 Internal Server Error.

Closes Issue #173

Signed-off-by: Eric Brown <browne@vmware.com>